Updating translations. Version 1.6.1!
git-svn-id: https://svn.code.sf.net/p/makoureactor/code/trunk@269 916e10b9-6b61-4450-83b7-7c74cc09f2a3

+++ core/field/BackgroundTexturesIO.cpp
@@ -174,7 +174,7 @@ bool BackgroundTexturesIOPS::read(BackgroundTexturesPS *textures)
 	textures->setHeaderEffect(headerEffect);
 	textures->setData(mimDataDec);
 
-	if (mimDataDec.size() != headerPalSize + headerImg.size + headerEffect.size) {
+	if (quint32(mimDataDec.size()) != headerPalSize + headerImg.size + headerEffect.size) {
 		qWarning() << "BackgroundTexturesIOPS::open padding after" << (mimDataDec.size() - int(headerPalSize + headerImg.size + headerEffect.size));
 	}
